+++ src/snowflake/snowpark/session.py
@@ -220,6 +220,7 @@ from snowflake.snowpark.types import (
     TimeType,
     VariantType,
     VectorType,
+    YearMonthIntervalType,
     FileType,
     _AtomicType,
 )
@@ -3685,6 +3686,7 @@ class Session:
                         TimestampType,
                         VariantType,
                         VectorType,
+                        YearMonthIntervalType,
                         FileType,
                     ),
                 )
@@ -3766,6 +3768,8 @@ class Session:
                     converted_row.append(value)
                 elif isinstance(data_type, VectorType):
                     converted_row.append(json.dumps(value, cls=PythonObjJSONEncoder))
+                elif isinstance(data_type, YearMonthIntervalType):
+                    converted_row.append(value)
                 else:
                     raise TypeError(
                         f"Cannot cast {type(value)}({value}) to {str(data_type)}."
@@ -3814,6 +3818,8 @@ class Session:
                 )
             elif isinstance(field.datatype, FileType):
                 project_columns.append(to_file(column(name)).as_(name))
+            elif isinstance(field.datatype, YearMonthIntervalType):
+                project_columns.append(column(name).cast(field.datatype).as_(name))
             else:
                 project_columns.append(column(name))