+++ CHANGELOG.md
@@ -41,6 +41,8 @@
   - `current_transaction`
   - `getbit`
 
+- Added a new datatype `YearMonthIntervalType` that allows users to create intervals for datetime operations.
+
 #### Bug Fixes
 
 - Fixed the repr of TimestampType to match the actual subtype it represents.
@@ -60,6 +62,7 @@
 ### Snowpark pandas API Updates
 
 #### New Features
+
 - Completed support for `pd.read_snowflake()`, `pd.to_iceberg()`,
   `pd.to_pandas()`, `pd.to_snowpark()`, `pd.to_snowflake()`,
   `DataFrame.to_iceberg()`, `DataFrame.to_pandas()`, `DataFrame.to_snowpark()`,
@@ -70,6 +73,7 @@
 - Added support for `Index.get_level_values()`.
 
 #### Improvements
+
 - Set the default transfer limit in hybrid execution for data leaving Snowflake to 100k, which can be overridden with the SnowflakePandasTransferThreshold environment variable. This configuration is appropriate for scenarios with two available engines, "Pandas" and "Snowflake" on relational workloads.
 - Improve import error message by adding '--upgrade' to 'pip install "snowflake-snowpark-python[modin]"' in the error message.
 - Reduce the telemetry messages from the modin client by pre-aggregating into 5 second windows and only keeping a narrow band of metrics which are useful for tracking hybrid execution and native pandas performance.
@@ -229,7 +233,6 @@
 - Added a ttl cache to describe queries. Repeated queries in a 15 second interval will use the cached value rather than requery Snowflake.
 - Added a parameter `fetch_with_process` to `DataFrameReader.dbapi` (PrPr) to enable multiprocessing for parallel data fetching in local ingestion. By default, local ingestion uses multithreading. Multiprocessing may improve performance for CPU-bound tasks like Parquet file generation.
 - Added a new function `snowflake.snowpark.functions.model` that allows users to call methods of a model.
-- Added a new datatype `YearMonthIntervalType` that allows users to create intervals for datetime operations.
 
 #### Improvements