+++ CHANGELOG.md
@@ -1,5 +1,31 @@
 # Release History
 
+## 1.39.0 (YYYY-MM-DD)
+
+### Snowpark Python API Updates
+
+- Added a new datatype `YearMonthIntervalType` that allows users to create intervals for datetime operations.
+
+#### New Features
+
+#### Bug Fixes
+
+#### Deprecations
+
+#### Dependency Updates
+
+#### Improvements
+
+### Snowpark pandas API Updates
+
+#### New Features
+
+#### Improvements
+
+#### Dependency Updates
+
+#### Bug Fixes
+
 ## 1.38.0 (YYYY-MM-DD)
 
 ### Snowpark Python API Updates
@@ -40,8 +66,6 @@
   - `current_secondary_roles`
   - `current_transaction`
   - `getbit`
-
-- Added a new datatype `YearMonthIntervalType` that allows users to create intervals for datetime operations.
 
 #### Bug Fixes